create useFarmer and new function in firebase
